Namespace CPS-specific Social Embed logic
