+++ src/app/containers/SocialEmbed/Cps/index.test.jsx
@@ -3,8 +3,8 @@ import { render } from '@testing-library/react';
 import { shouldMatchSnapshot } from '@bbc/psammead-test-helpers';
 import loggerMock from '#testHelpers/loggerMock';
 import { SOCIAL_EMBED_RENDERED } from '#lib/logger.const';
-import SocialEmbedContainer from '.';
-import withContexts from './testHelper';
+import CpsSocialEmbedContainer from '.';
+import withContexts from '../utilities/testHelper';
 import { twitterBlock, twitterBlockNoEmbed } from './fixtures';
 
 /* eslint-disable react/prop-types */
@@ -14,7 +14,7 @@ jest.mock('react-lazyload', () => {
   };
 });
 
-describe('SocialEmbedContainer', () => {
+describe('CpsSocialEmbedContainer', () => {
   afterEach(() => {
     loggerMock.info.mockClear();
   });
@@ -22,7 +22,7 @@ describe('SocialEmbedContainer', () => {
   describe('Canonical', () => {
     it('should render and unmount correctly', () => {
       const { container, unmount } = render(
-        withContexts(SocialEmbedContainer, {
+        withContexts(CpsSocialEmbedContainer, {
           isAmp: false,
           isEnabled: true,
         })({ blocks: [twitterBlock] }),
@@ -48,7 +48,7 @@ describe('SocialEmbedContainer', () => {
 
     it('should not render when disabled', () => {
       const { container } = render(
-        withContexts(SocialEmbedContainer, {
+        withContexts(CpsSocialEmbedContainer, {
           isAmp: false,
           isEnabled: false,
         })({ blocks: [twitterBlock] }),
@@ -59,7 +59,7 @@ describe('SocialEmbedContainer', () => {
 
     shouldMatchSnapshot(
       'should render correctly without an embed block',
-      withContexts(SocialEmbedContainer, {
+      withContexts(CpsSocialEmbedContainer, {
         isAmp: false,
         isEnabled: true,
       })({ blocks: [twitterBlockNoEmbed] }),
@@ -69,7 +69,7 @@ describe('SocialEmbedContainer', () => {
   describe('AMP', () => {
     shouldMatchSnapshot(
       'should render correctly',
-      withContexts(SocialEmbedContainer, {
+      withContexts(CpsSocialEmbedContainer, {
         isAmp: true,
         isEnabled: true,
       })({ blocks: [twitterBlock] }),
